#40 закончил мотодику просчетов EnergyFlow в сервисе - CreateTotalFluxDensityDto.cs модель для хранения суммарных подсчетов - PowerDensitySummation суммирование значений по всем передатчикам в проекте - EnergyResult.cs изменил тип переменной, дистанция целое число - IEnergyFlowService.cs изменил возвращаемую модель

+++ Domain/Entities/EnergyResult.cs
@@ -4,7 +4,7 @@ namespace Domain.Entities;
 
 public class EnergyResult : BaseEntity
 {
-    public decimal Distance { get; set; }
+    public int Distance { get; set; }
     public decimal Value { get; set; }
     public Guid AntennaTranslatorId { get; set; }
     public AntennaTranslator AntennaTranslator { get; set; }
